refactor(geo): extrai normalização de busca textual para helper compartilhado
Move RemoverDiacriticos e EscaparCuringasLike do CidadeReader para o helper
estático BuscaTextualNormalizada, que monta o padrão %termo% acento/caixa-
insensível sobre colunas *_normalizado. O CidadeReader passa a consumir o
helper, eliminando a duplicação antes de os readers de bairro e logradouro
reusarem a mesma lógica de busca.
