+++ src/geo/Unifesspa.UniPlus.Geo.Infrastructure/Persistence/Readers/CidadeReader.cs
@@ -1,9 +1,6 @@
 namespace Unifesspa.UniPlus.Geo.Infrastructure.Persistence.Readers;
 
 using System.Diagnostics.CodeAnalysis;
-using System.Globalization;
-using System.Text;
-
 using Microsoft.EntityFrameworkCore;
 
 using Unifesspa.UniPlus.Geo.Application.Abstractions;
@@ -62,12 +59,10 @@ internal sealed class CidadeReader : ICidadeReader
             // (a sobrecarga de 2 args do ILike emite ESCAPE '', desligando o \).
             // Guard NomeNormalizado != null: ILIKE NULL não casa — cidades sem
             // nome_normalizado simplesmente não aparecem na busca (aparecem sem filtro).
-            string termo = EscaparCuringasLike(RemoverDiacriticos(filtro.Termo!.Trim()));
-            string pattern = "%" + termo + "%";
-            const string escape = @"\";
+            string pattern = BuscaTextualNormalizada.CriarPadraoContem(filtro.Termo!);
             query = query.Where(c =>
                 c.NomeNormalizado != null &&
-                EF.Functions.ILike(c.NomeNormalizado, pattern, escape));
+                EF.Functions.ILike(c.NomeNormalizado, pattern, BuscaTextualNormalizada.Escape));
         }
 
         // Keyset bidirecional (ADR-0089) sobre a query JÁ filtrada — ordenação,
@@ -103,24 +98,4 @@ internal sealed class CidadeReader : ICidadeReader
             .ConfigureAwait(false);
     }
 
-    // Aplica a mesma transformação que produz nome_normalizado na fonte: decompõe em
-    // NFD e descarta NonSpacingMark (diacríticos). Normaliza o TERMO de busca em C#
-    // antes de montar o padrão ILIKE — a coluna já está sem acento, então não há
-    // chamada de DB function no lado do parâmetro. (Espelha UnidadeRepository.)
-    private static string RemoverDiacriticos(string texto)
-    {
-        string decomposto = texto.Normalize(NormalizationForm.FormD);
-        return new string([.. decomposto.Where(c =>
-            CharUnicodeInfo.GetUnicodeCategory(c) != UnicodeCategory.NonSpacingMark)]);
-    }
-
-    // Escapa os curingas do LIKE/ILIKE (\ % _) para que metacaracteres digitados
-    // pelo usuário sejam comparados como texto literal, não como wildcards — sem
-    // isso, q="_" ou q="%" casaria quase qualquer registro. A barra é escapada
-    // primeiro para não duplicar as inseridas adiante em % e _. (Espelha UnidadeRepository.)
-    private static string EscaparCuringasLike(string termo) =>
-        termo
-            .Replace("\\", "\\\\", StringComparison.Ordinal)
-            .Replace("%", "\\%", StringComparison.Ordinal)
-            .Replace("_", "\\_", StringComparison.Ordinal);
 }
